move mask_not_all_zeros back to common_methods_invocations, since it's used also elsewhere
